Reworked max iterations logic
Rather than having a hard coded max iterations value for all presets each preset now has its own max iterations value.
Now when users try to increase the iteration count of a system they cannot go above the max iteration count for the preset. This can only be changed in the text file for the preset and is done so at the users own risk
- Added Max Iterations to preset files
- Added max iterations to preset class
- Added max iterations to file reader
- Changed warnings in both parameter window and renderer
- Parameter window can now not increase iterations past the max count
- Parameter window now shows the max iteration count